Configuration Improvements
- Change game width and height from floats to integers.
- Pass command-line arg to filesystem functions (PhysFS WIP).
- Allow getting falling back if a config value doesn't exist.
- Move several debug configurations into the graphics section.
- Support the old debug graphic configurations as fallbacks.
- Improve formatting/style and consistently use Unix line endings.

+++ include/game.hpp
@@ -96,8 +96,8 @@ public:
     // Toggle fullscreen mode
     void set_fullscreen(bool fullscreen);
     // Game dimensions
-    float game_width(bool magnified = true) const;
-    float game_height(bool magnified = true) const;
+    int game_width(bool magnified = true) const;
+    int game_height(bool magnified = true) const;
     // Get screen magnification
     float get_magnification() const {
         return magnification;
